requirements.txt
google-generativeai
absl-py
